Changes in the run script
We found it more convenient to mount the data and source to the home directory. But it depends on what each user prefers. 
Moreover, we attached a network connection to monitor our experiments through wandb. If one is to use tensorboard, a port needs to be assigned and opened in the rooter settings.